+++ management/model_identity_provider_paypal.go
@@ -16,9 +16,11 @@ import (
 
 // IdentityProviderPaypal struct for IdentityProviderPaypal
 type IdentityProviderPaypal struct {
+	Links map[string]interface{} `json:"_links,omitempty"`
 	// The description of the IdP.
 	Description *string `json:"description,omitempty"`
-	Enabled EnumEnabledStatus `json:"enabled"`
+	// The current enabled state of the IdP.
+	Enabled bool `json:"enabled"`
 	Environment *ObjectEnvironment `json:"environment,omitempty"`
 	Icon *IdentityProviderCommonIcon `json:"icon,omitempty"`
 	// The resource ID.
@@ -28,6 +30,10 @@ type IdentityProviderPaypal struct {
 	Name string `json:"name"`
 	Registration *IdentityProviderCommonRegistration `json:"registration,omitempty"`
 	Type EnumIdentityProviderExt `json:"type"`
+	// The time the resource was created.
+	CreatedAt *string `json:"createdAt,omitempty"`
+	// The time the resource was last updated.
+	UpdatedAt *string `json:"updatedAt,omitempty"`
 	// A string that specifies the application ID from PayPal. This is a required property.
 	ClientId string `json:"clientId"`
 	// A string that specifies the application secret from PayPal. This is a required property.
@@ -40,7 +46,7 @@ type IdentityProviderPaypal struct {
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
 // will change when the set of required properties is changed
-func NewIdentityProviderPaypal(enabled EnumEnabledStatus, name string, type_ EnumIdentityProviderExt, clientId string, clientSecret string, clientEnvironment string) *IdentityProviderPaypal {
+func NewIdentityProviderPaypal(enabled bool, name string, type_ EnumIdentityProviderExt, clientId string, clientSecret string, clientEnvironment string) *IdentityProviderPaypal {
 	this := IdentityProviderPaypal{}
 	this.Enabled = enabled
 	this.Name = name
@@ -57,6 +63,38 @@ func NewIdentityProviderPaypal(enabled EnumEnabledStatus, name string, type_ Enu
 func NewIdentityProviderPaypalWithDefaults() *IdentityProviderPaypal {
 	this := IdentityProviderPaypal{}
 	return &this
+}
+
+// GetLinks returns the Links field value if set, zero value otherwise.
+func (o *IdentityProviderPaypal) GetLinks() map[string]interface{} {
+	if o == nil || o.Links == nil {
+		var ret map[string]interface{}
+		return ret
+	}
+	return o.Links
+}
+
+// GetLinksOk returns a tuple with the Links field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *IdentityProviderPaypal) GetLinksOk() (map[string]interface{}, bool) {
+	if o == nil || o.Links == nil {
+		return nil, false
+	}
+	return o.Links, true
+}
+
+// HasLinks returns a boolean if a field has been set.
+func (o *IdentityProviderPaypal) HasLinks() bool {
+	if o != nil && o.Links != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetLinks gets a reference to the given map[string]interface{} and assigns it to the Links field.
+func (o *IdentityProviderPaypal) SetLinks(v map[string]interface{}) {
+	o.Links = v
 }
 
 // GetDescription returns the Description field value if set, zero value otherwise.
@@ -92,9 +130,9 @@ func (o *IdentityProviderPaypal) SetDescription(v string) {
 }
 
 // GetEnabled returns the Enabled field value
-func (o *IdentityProviderPaypal) GetEnabled() EnumEnabledStatus {
+func (o *IdentityProviderPaypal) GetEnabled() bool {
 	if o == nil {
-		var ret EnumEnabledStatus
+		var ret bool
 		return ret
 	}
 
@@ -103,7 +141,7 @@ func (o *IdentityProviderPaypal) GetEnabled() EnumEnabledStatus {
 
 // GetEnabledOk returns a tuple with the Enabled field value
 // and a boolean to check if the value has been set.
-func (o *IdentityProviderPaypal) GetEnabledOk() (*EnumEnabledStatus, bool) {
+func (o *IdentityProviderPaypal) GetEnabledOk() (*bool, bool) {
 	if o == nil {
 		return nil, false
 	}
@@ -111,7 +149,7 @@ func (o *IdentityProviderPaypal) GetEnabledOk() (*EnumEnabledStatus, bool) {
 }
 
 // SetEnabled sets field value
-func (o *IdentityProviderPaypal) SetEnabled(v EnumEnabledStatus) {
+func (o *IdentityProviderPaypal) SetEnabled(v bool) {
 	o.Enabled = v
 }
 
@@ -323,6 +361,70 @@ func (o *IdentityProviderPaypal) SetType(v EnumIdentityProviderExt) {
 	o.Type = v
 }
 
+// GetCreatedAt returns the CreatedAt field value if set, zero value otherwise.
+func (o *IdentityProviderPaypal) GetCreatedAt() string {
+	if o == nil || o.CreatedAt == nil {
+		var ret string
+		return ret
+	}
+	return *o.CreatedAt
+}
+
+// GetCreatedAtOk returns a tuple with the CreatedAt field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *IdentityProviderPaypal) GetCreatedAtOk() (*string, bool) {
+	if o == nil || o.CreatedAt == nil {
+		return nil, false
+	}
+	return o.CreatedAt, true
+}
+
+// HasCreatedAt returns a boolean if a field has been set.
+func (o *IdentityProviderPaypal) HasCreatedAt() bool {
+	if o != nil && o.CreatedAt != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetCreatedAt gets a reference to the given string and assigns it to the CreatedAt field.
+func (o *IdentityProviderPaypal) SetCreatedAt(v string) {
+	o.CreatedAt = &v
+}
+
+// GetUpdatedAt returns the UpdatedAt field value if set, zero value otherwise.
+func (o *IdentityProviderPaypal) GetUpdatedAt() string {
+	if o == nil || o.UpdatedAt == nil {
+		var ret string
+		return ret
+	}
+	return *o.UpdatedAt
+}
+
+// GetUpdatedAtOk returns a tuple with the UpdatedAt field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *IdentityProviderPaypal) GetUpdatedAtOk() (*string, bool) {
+	if o == nil || o.UpdatedAt == nil {
+		return nil, false
+	}
+	return o.UpdatedAt, true
+}
+
+// HasUpdatedAt returns a boolean if a field has been set.
+func (o *IdentityProviderPaypal) HasUpdatedAt() bool {
+	if o != nil && o.UpdatedAt != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetUpdatedAt gets a reference to the given string and assigns it to the UpdatedAt field.
+func (o *IdentityProviderPaypal) SetUpdatedAt(v string) {
+	o.UpdatedAt = &v
+}
+
 // GetClientId returns the ClientId field value
 func (o *IdentityProviderPaypal) GetClientId() string {
 	if o == nil {
@@ -397,6 +499,9 @@ func (o *IdentityProviderPaypal) SetClientEnvironment(v string) {
 
 func (o IdentityProviderPaypal) MarshalJSON() ([]byte, error) {
 	toSerialize := map[string]interface{}{}
+	if o.Links != nil {
+		toSerialize["_links"] = o.Links
+	}
 	if o.Description != nil {
 		toSerialize["description"] = o.Description
 	}
@@ -423,6 +528,12 @@ func (o IdentityProviderPaypal) MarshalJSON() ([]byte, error) {
 	}
 	if true {
 		toSerialize["type"] = o.Type
+	}
+	if o.CreatedAt != nil {
+		toSerialize["createdAt"] = o.CreatedAt
+	}
+	if o.UpdatedAt != nil {
+		toSerialize["updatedAt"] = o.UpdatedAt
 	}
 	if true {
 		toSerialize["clientId"] = o.ClientId